Rename job email process
